test<refactor>: centralizing generation of faker properties
